Move mobile compose prompt to inner component
